Add the start of new logic for ProjectList feature
I'm trying to create an effect where any given ProjectList can have a
set maximum number of rows displayed, which reacts to the dynamic
nature of the grid container for the ProjectPanel components. For
example, if the default is 2 rows, when the containing element for the
ProjectList is wide enough that 3 columns can be displayed, this means
6 ProjectPanels will be displayed. If the user then shrinks the window
down to 2 columns, only 4 will be displayed. The user should be able to
click a 'Show More' button to increase the number of rows and therefore
the number of displayed ProjectPanels in the grid, if any. So far, I
have achieved making the component re-render when the number of columns
changes, by memoising and accessing the element's computed styles.