Refactor Projects
Moved the ProjectList into its own component for reusability. Also
standardised the parameters for a ProjectPanel in its own file:
ProjectPanelParams.js, from which the exported object can be used to
create consistent propTypes and defaultProps across the relevant
components in this feature, as we only need to change it in one place
now.