Create scrolling functionality for Nav buttons
The Navigation buttons at the top of the landing page will now call
the 'scrollIntoView' Element method when clicked to position the
viewport around the specified element. The 'navButtonClicked' function
in the Landing page component calculates whether or not the element's
height, plus a specified 'buffer' added to the top and bottom of the
element's height, is larger or smaller than the viewport height, and
as a result will scroll to the top or centre of the element,
respectively.