Change fade-in animations for elements
Some features on the Landing page have a fade-in animation that, until
now, had been playing immediately as the page was rendered. By adding
an eventListener to the window object that checks for scroll events, I
have made it so the wrapper element for the components that employ this
animation have the "animate" attribute added to it, which I have used to
delay the start of the animation only when the wrapper is scrolled into
view/the page loads with it already in view.