Refactor technologiesInformation utility function
The validateTechnologies function can now accept 'null' as its first
parameter in-place of a category string and instead of checking whether
that specific technology is part of a category, it will just check
whether it is defined at all in ANY of the categories. It now also takes
an optional third argument which, if specified, will be displayed as
the component's name that called the function in case of a validation
error.